feat: remove ds test from main website using env
