+++ .env.example
@@ -1,0 +1,1 @@
+NEXT_PUBLIC_IS_DEVELOPMENT=true
