+++ src/app/page.tsx
@@ -1,44 +1,17 @@
-"use client";
-import { Button } from "@/components/ui/button";
-import Chip from "@/components/ui/chip";
-import Countdown from "@/components/elements/Countdown";
-import Pagination from "@/components/ui/pagination";
-import Tooltip from "@/components/ui/tooltip";
-import { useState } from "react";
-import { ContactPerson } from "@/components/elements/ContactPerson";
-import TestimonyCarousel from "@/components/elements/testimony/TestimonyCarousel";
-import { betisTestimoni, betisTestimoniOptions } from "@/components/elements/testimony/const";
-
-export default function Home() {
-  const [page, setPage] = useState(1);
+export default function Page() {
   return (
-    <div className="h-[400vh] flex justify-center flex-col gap-6 my-10 items-center">
-      <div className="grid grid-cols-1 gap-6">
-        <Button>Click me</Button>
-        <Button isLoading>Click me</Button>
-        <Button variant="secondary">Click me</Button>
-        <Button isLoading variant="secondary">
-          Click me
-        </Button>
-        <Button variant="tertiary">Click me</Button>
-        <Button isLoading variant="tertiary">
-          Click me
-        </Button>
-        <Button variant="destructive">Click me</Button>
-        <Button isLoading variant="destructive">
-          Click me
-        </Button>
+    <main className="min-h-screen flex flex-col justify-center items-center text-center">
+      <h1 className="font-cinzel text-[96px] text-violet-900/40">
+        Coming Soon
+      </h1>
+      <div className="">
+        <h2 className="font-cinzel text-t6 md:text-t3">
+          We are cooking something good!
+        </h2>
+        <h3 className="font-raleway text-t8 md:text-t6 max-md:mt-3">
+          Please wait and check back later.
+        </h3>
       </div>
-      <Countdown displayDate targetDate={new Date("2024-12-31T23:59:59")} />
-      <Chip>Chip</Chip>
-      <Chip variant="secondary">Chip</Chip>
-      <Chip variant="tertiary">Chip</Chip>
-      <Button>
-        <Tooltip text="Tooltip">Hover me</Tooltip>
-      </Button>
-      <Pagination currentPage={page} totalPages={10} onPageChange={setPage} />
-      <ContactPerson />
-      <TestimonyCarousel slides={betisTestimoni} options={betisTestimoniOptions} />
-    </div>
+    </main>
   );
-}
+};
